Django Rest Framework/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==3.0.7
djangorestframework==3.14.0
importlib-metadata==6.0.0
Markdown==3.4.1
pycodestyle==2.10.0
Pygments==2.14.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
zipp==3.15.0
